move memory/detail to memory/allocation
