+++ paddle/fluid/pybind/CMakeLists.txt
@@ -83,11 +83,6 @@ if(WITH_CNCL)
 endif()
 
 if(NOT WIN32)
-  set(PYBIND_DEPS ${PYBIND_DEPS} data_loader)
-  set(PYBIND_DEPS ${PYBIND_DEPS} mmap_allocator)
-  if(WITH_GPU)
-    set(PYBIND_DEPS ${PYBIND_DEPS} cuda_ipc_allocator)
-  endif()
   if(WITH_NCCL OR WITH_RCCL)
     set(PYBIND_DEPS ${PYBIND_DEPS} nccl_context)
     set(PYBIND_DEPS ${PYBIND_DEPS} heter_ccl_context)
